[FLINK-3209] Remove Unused ProcessingTime, EventTime and AbstractTime
Only keep Time for specifying time durations/intervals.
