+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/api/windowing/evictors/TimeEvictor.java
@@ -19,7 +19,7 @@ package org.apache.flink.streaming.api.windowing.evictors;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.Iterables;
-import org.apache.flink.streaming.api.windowing.time.AbstractTime;
+import org.apache.flink.streaming.api.windowing.time.Time;
 import org.apache.flink.streaming.api.windowing.windows.Window;
 import org.apache.flink.streaming.runtime.streamrecord.StreamRecord;
 
@@ -67,7 +67,7 @@ public class TimeEvictor<W extends Window> implements Evictor<Object, W> {
 	 *
 	 * @param windowSize The amount of time for which to keep elements.
 	 */
-	public static <W extends Window> TimeEvictor<W> of(AbstractTime windowSize) {
+	public static <W extends Window> TimeEvictor<W> of(Time windowSize) {
 		return new TimeEvictor<>(windowSize.toMilliseconds());
 	}
 }
