+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/api/windowing/time/Time.java
@@ -18,32 +18,55 @@
 
 package org.apache.flink.streaming.api.windowing.time;
 
-import org.apache.flink.streaming.api.TimeCharacteristic;
-
 import java.util.concurrent.TimeUnit;
+
+import static com.google.common.base.Preconditions.checkNotNull;
 
 /**
  * The definition of a time interval for windowing. The time characteristic referred
  * to is the default time characteristic set on the execution environment.
  */
-public final class Time extends AbstractTime {
+public final class Time {
+
+	/** The time unit for this policy's time interval */
+	private final TimeUnit unit;
+
+	/** The size of the windows generated by this policy */
+	private final long size;
 
 	/** Instantiation only via factory method */
 	private Time(long size, TimeUnit unit) {
-		super(size, unit);
+		this.unit = checkNotNull(unit, "time unit may not be null");
+		this.size = size;
+
 	}
 
-	@Override
-	public AbstractTime makeSpecificBasedOnTimeCharacteristic(TimeCharacteristic timeCharacteristic) {
-		switch (timeCharacteristic) {
-			case ProcessingTime:
-				return ProcessingTime.of(getSize(), getUnit());
-			case IngestionTime:
-			case EventTime:
-				return EventTime.of(getSize(), getUnit());
-			default:
-				throw new IllegalArgumentException("Unknown time characteristic");
-		}
+	// ------------------------------------------------------------------------
+	//  Properties
+	// ------------------------------------------------------------------------
+
+	/**
+	 * Gets the time unit for this policy's time interval.
+	 * @return The time unit for this policy's time interval.
+	 */
+	public TimeUnit getUnit() {
+		return unit;
+	}
+
+	/**
+	 * Gets the length of this policy's time interval.
+	 * @return The length of this policy's time interval.
+	 */
+	public long getSize() {
+		return size;
+	}
+
+	/**
+	 * Converts the time interval to milliseconds.
+	 * @return The time interval in milliseconds.
+	 */
+	public long toMilliseconds() {
+		return unit.toMillis(size);
 	}
 
 	// ------------------------------------------------------------------------
